chore: remove obsolete debug configuration and scripts, and add new mygit implementation for Git functionality
