--- debug/codecrafters.yml
@@ -1,11 +1,0 @@
-# Set this to true if you want debug logs.
-#
-# These can be VERY verbose, so we suggest turning them off
-# unless you really need them.
-debug: false
-
-# Use this to change the Go version used to run your code
-# on Codecrafters.
-#
-# Available versions: go-1.24
-language_pack: go-1.24
