--- debug/your_program.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-go mod tidy
-go build -o mygit ./cmd/mygit
-exec ./mygit "$@"
