+++ internal/test_helpers/pass_all/codecrafters.yml
@@ -1,11 +1,11 @@
-# This is the current stage that you're on.
-#
-# Whenever you want to advance to the next stage,
-# bump this to the next number.
-current_stage: 1
-
 # Set this to true if you want debug logs.
 #
 # These can be VERY verbose, so we suggest turning them off
 # unless you really need them.
-debug: true
+debug: false
+
+# Use this to change the Go version used to run your code
+# on Codecrafters.
+#
+# Available versions: go-1.24
+language_pack: go-1.24
