--- internal/test_helpers/pass_all/your_git.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-
-if [ "$1" = "write-tree" ]
-then
-  git add .
-fi
-
-exec git "$@"
