+++ internal/test_helpers/pass_all/your_program.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+exec $(dirname $0)/mygit "$@"
